+++ src/test/groovy/org/m2ci/msp/flaml/FlamlPluginFunctionalTest.groovy
@@ -2,7 +2,7 @@ package org.m2ci.msp.flaml
 
 import org.gradle.testkit.runner.GradleRunner
 import org.gradle.testkit.runner.TaskOutcome
-import org.testng.annotations.BeforeClass
+import org.testng.annotations.BeforeGroups
 import org.testng.annotations.DataProvider
 import org.testng.annotations.Test
 
@@ -10,17 +10,38 @@ class FlamlPluginFunctionalTest {
 
     GradleRunner gradle
 
-    @BeforeClass
-    void setup() {
+    void setupGradleAndProjectDir(String buildScriptResourceName, String... resourceNames) {
         def projectDir = File.createTempDir()
-        gradle = GradleRunner.create().withPluginClasspath().withProjectDir(projectDir)
-        ['build-core.gradle', 'build-extraction.gradle', 'build-generation.gradle',
-         'foobar.flac', 'foobar.yaml', 'foobar.wav', 'foobar.TextGrid', 'foo_padded.wav', 'baz_padded.wav',
-         'foo.wav', 'bar.wav', 'baz.wav', 'foo.lab', 'baz.lab', 'foo.txt', 'baz.txt'].each { resourceName ->
+        new File(projectDir, 'settings.gradle').createNewFile()
+        new File(projectDir, 'build.gradle').withWriter { writer ->
+            writer << this.class.getResourceAsStream(buildScriptResourceName)
+        }
+        resourceNames.each { resourceName ->
             new File(projectDir, resourceName).withOutputStream { stream ->
-                stream << this.getClass().getResourceAsStream(resourceName)
+                stream << this.class.getResourceAsStream(resourceName)
             }
         }
+        gradle = GradleRunner.create().withPluginClasspath().withProjectDir(projectDir).forwardOutput()
+    }
+
+    @BeforeGroups(groups = 'core')
+    void setupCore() {
+        setupGradleAndProjectDir('build-core.gradle',
+                'foobar.flac', 'foobar.yaml')
+    }
+
+    @BeforeGroups(groups = 'extraction')
+    void setupExtraction() {
+        setupGradleAndProjectDir('build-extraction.gradle',
+                'foobar.flac', 'foobar.yaml',
+                'foobar.TextGrid', 'foo.wav', 'bar.wav', 'baz.wav', 'foo.lab', 'baz.lab', 'foo.txt', 'baz.txt')
+    }
+
+    @BeforeGroups(groups = 'generation')
+    void setupGeneration() {
+        setupGradleAndProjectDir('build-generation.gradle',
+                'foobar.flac', 'foobar.yaml',
+                'foo_padded.wav', 'baz_padded.wav', 'bar.wav')
     }
 
     @DataProvider
@@ -31,8 +52,7 @@ class FlamlPluginFunctionalTest {
                 ['tasks'],
                 ['hasTaskDescriptions'],
                 ['hasExtension'],
-                ['hasFlamlResources'],
-                ['hasTestResources'],
+                ['hasFlamlResources']
         ]
     }
 
@@ -57,24 +77,23 @@ class FlamlPluginFunctionalTest {
         ]
     }
 
-    @Test(dataProvider = 'coreTasks')
+    void runGradleTask(String taskName) {
+        def result = gradle.withArguments('--warning-mode', 'all', taskName).build()
+        assert result.task(":$taskName").outcome in [TaskOutcome.SUCCESS, TaskOutcome.UP_TO_DATE]
+    }
+
+    @Test(groups = 'core', dataProvider = 'coreTasks')
     void testCoreTasks(String taskName) {
-        def result = gradle.withArguments('--build-file', 'build-core.gradle', taskName).build()
-        println result.output
-        assert result.task(":$taskName").outcome in [TaskOutcome.SUCCESS, TaskOutcome.UP_TO_DATE]
+        runGradleTask(taskName)
     }
 
-    @Test(dataProvider = 'extractionTasks')
+    @Test(groups = 'extraction', dataProvider = 'extractionTasks')
     void testExtractionTasks(String taskName) {
-        def result = gradle.withArguments('--build-file', 'build-extraction.gradle', taskName).build()
-        println result.output
-        assert result.task(":$taskName").outcome in [TaskOutcome.SUCCESS, TaskOutcome.UP_TO_DATE]
+        runGradleTask(taskName)
     }
 
-    @Test(dataProvider = 'generationTasks')
+    @Test(groups = 'generation', dataProvider = 'generationTasks')
     void testGenerationTasks(String taskName) {
-        def result = gradle.withArguments('--build-file', 'build-generation.gradle', taskName).build()
-        println result.output
-        assert result.task(":$taskName").outcome in [TaskOutcome.SUCCESS, TaskOutcome.UP_TO_DATE]
+        runGradleTask(taskName)
     }
 }